Add GtkClutterActor for embedding Gtk+ widget in clutter
GtkClutterActor is the container for gtk+ widgets. It has a
GtkClutterOfffscreen widget that you can get and put other stuff in. This
gets drawn to an offscreen windows, and that pixmap is put into a
ClutterTexture object that is a child of the GtkClutterActor.

In order for the widget hierarchy to be correct GtkClutterEmbed is made
a container of GtkClutterOffscreen widgets.