GtkClutterStandinBin: a private container Clutter-side for GtkClutterStandin
This actor inherits ClutterGroup to provide a single-widget bin (similar to
GtkBin). This allows us to track when the child actor's size has changed, and
trigger a resize in GTK+.

It also will permit us to change the actor for a GtkStandinBin after
construction (not implemented yet).

This change has triggered a reparenting bug as highlighted by reparent-test.c.