Avoid an allocation cycle in StandinBin
The StandinBin gets the allocation from GTK+ and pushes it to the Actor
it is holding. This is done by calling allocate() out of the blue inside
the GTK+ size-allocate implementation and then calling
get_allocation_box() inside the Clutter allocate implementation. Works
like a charm.

Except that calling allocate() out of the blue is wrong, and calling
get_allocation_box() in the middle of an allocation is going to cause
a reallocation cycle. Both things are really bad.

If you are a little fuzzy on the whole good/bad thing then, to put it to
scale, this is not the case where all life as you know it stops
instantaneously and every molecule in your body explodes at the speed of
light, but it gets close.

Instead, we should store the GTK+ allocation in size-allocate and set it
when we get asked to be allocated by Clutter in allocate.