embed: Update input devices
The event handling in Clutter has been changed internally to always use
input devices to store the state of the pointers and to emit ENTER/LEAVE
event pairs consistently. This means that GtkClutterEmbed should be
responsible of updating the state of the input devices, since we're
completely overriding the event handling.