actor: Avoid layout cycles/2
We use a custom ::check-resize in GtkClutterOffscreen to protect
ourselves from containers that call queue_resize() - and queue a
resize of the Clutter scene along with the gtk+ hierarchy.

The ::check-resize in GtkClutterOffscreen is being called from within
the GtkClutterActor::allocate implementation during the initial set
up of the scene, through gtk_widget_size_allocate(). The ::check-resize
logic is useless in this case because we're in the middle of a relayout
anyway.