Backport widget embedding for gtk+ 2.x
Widget embedding in clutter-gtk was originally written for gtk+ 2.x; for
clutter-gtk 1.0 it was decided to move over to gtk+ 3 and clutter 1.6,
but it should still be possible to provide widget embedding in a gtk+
2.x environment. Hence, clutter-gtk 0.12 is the last release of
clutter-gtk tracking gtk+ 2.x.

Widget embedding has been backported from the master branch, and it's
using the gtk+/gdk 2.x API instead of the 3.x one.