embed: Set up a better initial state
The GtkClutterEmbed widget should turn on the APP_PAINTABLE flag, since
it's responsible for painting everything, including the background.

It should also set the h-expand and v-expand properties to TRUE, as the
container of the widget should allocate all the extra space; this allows
us to get rid of the explicit set_size_request() in a bunch of places.