embed: queue a draw on actors when embed state flag change
When we receive a state-flags-changed signal, we should queue a draw on
all of our children, since their appearance might be depending on
inherited style values that changed.

This should really be done generically by GTK, but doing it there is
very bad for performance with the current GTK code, so we do it here.