Add explicit subclassing notes for GtkClutterEmbed
The reason why GtkClutterEmbed is a GtkContainer is that we need to
provide an uninterrupted parents chain from the top-level to widgets
embedded using GtkClutterActor.

If you override GtkContainer's implementation when subclassing
GtkClutterEmbed you can only chain up - and end up triggering an
assertion failure, because GtkClutterEmbed only accepts children of
GtkClutterOffscreen type; GtkClutterOffscreen is an internal type,
though, so you can't use that knowlegde as a filter in your own code.

It's better to cut our losses, and just explicitly note in the
documentation that overriding GtkContainer is just A Bad Idea™ when
subclassing GtkClutterEmbed.