actor: Hide children before ourselves
Fixes errors like:
Gtk-WARNING **: GtkClutterEmbed 0x11a3300 is mapped but visible child GtkClutterOffscreen 0x15d0490 is not mapped
when destroying a GtkClutterActor.

GTK+'s invariants checks expect children to not be visible anymore
if they're not mapped. This can happen if we don't forcefully hide the
GtkClutterActor's widget before unrealizing it.

https://bugzilla.gnome.org/show_bug.cgi?id=675867