Make GtkClutterEmbed redraw resizes nicer
Currently when you vertically resize a GtkClutterEmbed we draw twice, once
with the old window size and once again when the ConfigureNotify gets
back with the right size. The first draw is at the wrong position due
to the wrong window height used to calculate the viewport position.

We fix this by immediately telling cogl about the new size by faking
a ConfigureNotify event.

https://bugzilla.gnome.org/show_bug.cgi?id=677493