Provide a fallback for embedding widgets into actors
Currently, we can only embed a GtkWidget into a ClutterActor on X11 by
using a Pixmap and the GLX texture_from_pixmap extension. There is no
reason why we shouldn't be able to allow the same to happen by using
Cairo surfaces — except the performance implications of making copies of
the contents of the Drawable used by GDK.

This should allow embedding widgets on platforms that do not have the
texture_from_pixmap extension.