wayland: Use the subsurface protocol to implement GtkClutterEmbed
The subsurface protocol lets us "embed" one surface within another. The
compositor will compose the two surfaces together to create a single
window. We use it here to take the surface we get from Clutter and make
that into a subsurface and then associate that subsurface with the main
surface coming from GTK+.

https://bugzilla.gnome.org/show_bug.cgi?id=695737