embed: rework foreign stage setup to work on gdk/wayland
On hide/unmap the GDK wayland backend might decide to destroy its
wayland surface [1]. There is no event notifying Clutter/Cogl of this.

We need to notify Clutter/Cogl that the surface has changed when the
embed widget becomes visible/invisible. So instead of mapping the
different mapped/visible/realized properties of the GtkWidget to the
ClutterStage, we take a different approach by considering that an
invisible GtkWidget means an unrealized ClutterStage.

[1] : https://git.gnome.org/browse/gtk+/commit/?id=87e2a7d4b23e633d0f2263d6ebd205c113eede0f

https://bugzilla.gnome.org/show_bug.cgi?id=754891